requirements.txt
PySimpleGUI==4.55.1
pytube==15.0.0
pillow==9.5.0
requests==2.28.2

